BREAKING CHANGE: Revert having collect return subclasses
This broke the supported use case if injecting the `collect` function in to any other component.
It broke because the `getMetadata( this )` call returned whatever component the `collect`
function was injected into.

The official way to subclass `Collection` will include overloading the `collect` method
to return an instance of the subclass instead of the base Collection.